finished final packages and optmised duke main class
